ENH: do not leave a gap for top interface if the top z distance is 0
Previously when the overhang is a slop, it's controled by the xy
distance. Now we control it with top z distance as long as it's an
interface.

Note: Not working with Snug support, which has a different mechanism of
support interface.

Jira: STUDIO-3991

Change-Id: I1b7c4cf892b4c4ec2f3e668bd966ab0971212364
(cherry picked from commit c094b5ec83dc98dca7a8b8c0438b8f1994134db6)