ENH: Improved safe position calculation for delayed photography
In the safe position calculation for time-lapse photography in traditional mode of piece-by-piece printing, for parts not currently being printed, the maximum expansion radius is used for parts higher than the current layer, and half of the expansion radius is used for parts lower than the current layer height; half of the expansion radius is used for the parts themselves. This reduces unnecessary avoidance.

jira: STUDIO-13194
Change-Id: Id6ce03af29c9ebba974ea57ec41cca757067027a